+++ lib/volt/models/model.rb
@@ -4,9 +4,14 @@ require 'volt/models/model_helpers'
 require 'volt/models/model_hash_behaviour'
 require 'volt/models/validations'
 require 'volt/models/model_state'
+require 'volt/models/modes'
 require 'volt/models/buffer'
 require 'volt/models/field_helpers'
 require 'volt/reactive/reactive_hash'
+require 'volt/models/validators/user_validation'
+require 'volt/models/dirty'
+require 'volt/reactive/class_eventable'
+require 'thread'
 
 module Volt
   class NilMethodCall < NoMethodError
@@ -25,9 +30,12 @@ module Volt
     include ModelState
     include Buffer
     include FieldHelpers
+    include UserValidatorHelpers
+    include Dirty
+    include ClassEventable
+    include Modes
 
-    attr_reader :attributes
-    attr_reader :parent, :path, :persistor, :options
+    attr_reader :attributes, :parent, :path, :persistor, :options
 
     INVALID_FIELD_NAMES = {
       :attributes => true,
@@ -42,6 +50,8 @@ module Volt
       @size_dep    = Dependency.new
       self.options = options
 
+      @new = (initial_state != :loaded)
+
       send(:attributes=, attributes, true)
 
       # Models start in a loaded state since they are normally setup from an
@@ -49,6 +59,9 @@ module Volt
       @state = :loaded
 
       @persistor.loaded(initial_state) if @persistor
+
+      # Trigger the new event, pass in :new
+      trigger!(:new, :new)
     end
 
     # the id is stored in a field named _id, so we setup _id to proxy to this
@@ -58,6 +71,11 @@ module Volt
 
     def _id=(val)
       self.__id = val
+    end
+
+    # Return true if the model hasn't been saved yet
+    def new?
+      @new
     end
 
     # Update the options
@@ -79,8 +97,8 @@ module Volt
         # Assign id first
         id = attrs.delete(:_id)
 
-        # When doing a mass-assign, we don't save until the end.
-        Model.nosave do
+        # When doing a mass-assign, we don't validate or save until the end.
+        Model.no_validate do
           self._id = id if id
 
           # Assign each attribute using setters
@@ -102,14 +120,12 @@ module Volt
       @deps.changed_all!
       @deps = HashDependency.new
 
-      unless initial_setup
-
-        # Let the persistor know something changed
-        if @persistor
-          # the changed method on a persistor should return a promise that will
-          # be resolved when the save is complete, or fail with a hash of errors.
-          return @persistor.changed
-        end
+      # Save the changes
+      if initial_setup
+        # Run initial validation
+        validate!
+      else
+        run_changed
       end
     end
 
@@ -133,12 +149,11 @@ module Volt
 
     def method_missing(method_name, *args, &block)
       if method_name[0] == '_'
-
         # Remove underscore
         method_name = method_name[1..-1]
         if method_name[-1] == '='
           # Assigning an attribute without the =
-          assign_attribute(method_name[0..-2], *args, &block)
+          assign_attribute(method_name[0..-2], args[0], &block)
         else
           read_attribute(method_name)
         end
@@ -149,19 +164,21 @@ module Volt
     end
 
     # Do the assignment to a model and trigger a changed event
-    def assign_attribute(method_name, *args, &block)
+    def assign_attribute(attribute_name, value, &block)
       self.expand!
       # Assign, without the =
-      attribute_name = method_name.to_sym
+      attribute_name = attribute_name.to_sym
 
       check_valid_field_name(attribute_name)
-
-      value = args[0]
 
       old_value = @attributes[attribute_name]
       new_value = wrap_value(value, [attribute_name])
 
       if old_value != new_value
+        # Track the old value
+        attribute_will_change!(attribute_name, old_value)
+
+        # Assign the new value
         @attributes[attribute_name] = new_value
 
         @deps.changed!(attribute_name)
@@ -172,13 +189,10 @@ module Volt
 
         # TODO: Can we make this so it doesn't need to be handled for non store collections
         # (maybe move it to persistor, though thats weird since buffers don't have a persistor)
-        clear_server_errors(attribute_name) if @server_errors
+        clear_server_errors(attribute_name) if @server_errors.present?
 
-        # Don't save right now if we're in a nosave block
-        if !defined?(Thread) || !Thread.current['nosave']
-          # Let the persistor know something changed
-          @persistor.changed(attribute_name) if @persistor
-        end
+        # Save the changes
+        run_changed(attribute_name)
       end
     end
 
@@ -290,25 +304,17 @@ module Volt
       "<#{self.class}:#{object_id} #{attributes.inspect}>"
     end
 
-    # Takes a block that when run, changes to models will not save inside of
-    if RUBY_PLATFORM == 'opal'
-      # Temporary stub for no save on client
-      def self.nosave
-        yield
-      end
-    else
-      def self.nosave
-        previous = Thread.current['nosave']
-        Thread.current['nosave'] = true
-        begin
-          yield
-        ensure
-          Thread.current['nosave'] = previous
-        end
-      end
+    def self.no_save(&block)
+      run_in_mode(:no_save, &block)
     end
 
+
     private
+
+    # no_validate mode should only be used internally
+    def self.no_validate(&block)
+      run_in_mode(:no_validate, &block)
+    end
 
     # Volt provides a few access methods to get more data about the model,
     # we want to prevent these from being assigned or accessed through
@@ -329,6 +335,52 @@ module Volt
       if persistor
         @persistor = persistor.new(self)
       end
+    end
+
+    # Called when something in the model changes.  Saves
+    # the model if there is a persistor, and changes the
+    # model to not be new.
+    #
+    # @return [Promise|nil] a promise for when the save is
+    #         complete
+    def run_changed(attribute_name=nil)
+      result = nil
+
+      # no_validate mode should only be used internally.  no_validate mode is a
+      # performance optimization that prevents validation from running after each
+      # change when assigning multile attributes.
+      unless in_mode?(:no_validate)
+        # Run the validations for all fields
+        validate!
+
+        # Buffers are allowed to be in an invalid state
+        unless buffer?
+          # First check that all local validations pass
+          if error_in_changed_attributes?
+            # Some errors are present, revert changes
+            revert_changes!
+
+            # After we revert, we need to validate again to get the error messages back
+            # TODO: Could probably cache the previous errors.
+            validate!
+          else
+            # No errors, tell the persistor to handle the change (usually save)
+
+            # Don't save right now if we're in a nosave block
+            unless in_mode?(:no_save)
+              # the changed method on a persistor should return a promise that will
+              # be resolved when the save is complete, or fail with a hash of errors.
+              result = @persistor.changed(attribute_name) if @persistor
+              @new = false
+
+              # Clear the change tracking
+              clear_tracked_changes!
+            end
+          end
+        end
+      end
+
+      return result
     end
   end
 end