+++ lib/volt/server/html_parser/view_parser.rb
@@ -1,6 +1,7 @@
 require 'volt/server/html_parser/sandlebars_parser'
 require 'volt/server/html_parser/view_scope'
 require 'volt/server/html_parser/if_view_scope'
+require 'volt/server/html_parser/component_view_scope'
 require 'volt/server/html_parser/view_handler'
 require 'volt/server/html_parser/each_scope'
 require 'volt/server/html_parser/textarea_scope'